add functionality discounts and change in companies functionality

+++ src/useCases/discounts.use.js
@@ -1,24 +1,41 @@
-import { Discounts } from "../models/discounts.models.js";
+import { Discounts } from "../models/discounts.model.js";
 import { StatusHttp } from "../libs/statusHttp.js";
 
-function getAll() {
-  return Discounts.find({}).populate("company");
+async function getAll() {
+  return await Discounts.find({}).populate("company");
 }
 
-function getById(id) {
-  return Discounts.findById(id).populate("company");
+async function getById(id) {
+  const discountFind = await Discounts.findById(id).populate("company");
+  if (discountFind) {
+    return discountFind;
+  } else {
+    throw new StatusHttp("Descuento no encontrado");
+  }
 }
 
-function deleteById(id) {
-  return Discounts.findByIdAndDelete(id);
+async function deleteById(id, company) {
+  const discount = await Discounts.findById(id).populate("company");
+  if (discount.company._id == company) {
+    return await Discounts.findByIdAndDelete(id);
+  } else {
+    throw new StatusHttp("No puedes eliminar un descuento que no te pertenece");
+  }
 }
 
 async function create(newDiscount, company) {
-  return Discounts.create(newDiscount, company);
+  return await Discounts.create({ ...newDiscount, company: company });
 }
 
-async function updated(idDiscount, updatedCompany) {
-  return Discounts.findByIdAndUpdate(idDiscount, updatedCompany);
+async function updated(idDiscount, updatedCompany, company) {
+  const discount = await Discounts.findById(idDiscount).populate("company");
+  if (discount.company._id == company) {
+    return await Discounts.findByIdAndUpdate(idDiscount, updatedCompany);
+  } else {
+    throw new StatusHttp(
+      "No puedes actualizar un descuento que no te pertenece"
+    );
+  }
 }
 
 export { getAll, getById, deleteById, updated, create };
